Extracted encodeQueryParameters from exposed utility functions and remove decoding check. The method is not supposed to be exposed along with the utility functions. ExecuteRequestUtils has been created to contain it. When encoding the query parameters, we do not check if they are already encoded as it is the responsibility of the workflow developer to provide a not encoded url.
